pyproject.toml
[project]
name = "paper-survey-agent"
version = "0.1.0"
description = "Agentic ML paper discovery pipeline that builds an Obsidian knowledge base"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.12",
    "aiohttp>=3.9",
    "beautifulsoup4>=4.12",
    "pyyaml>=6.0",
    "requests>=2.31",
    "arxiv>=2.0",
    "semanticscholar>=0.8",
]

[project.scripts]
paper-survey-agent = "paper_discovery.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
